+++ app/src/organisms/ErrorRecoveryFlows/utils/__tests__/getErrorKind.test.ts
@@ -1,0 +1,12 @@
+import { describe, expect, it } from 'vitest'
+
+import { ERROR_KINDS } from '../../constants'
+import { getErrorKind } from '../getErrorKind'
+
+describe('getErrorKind', () => {
+  it(`returns ${ERROR_KINDS.GENERAL_ERROR} if the errorType isn't handled explicitly`, () => {
+    const mockErrorType = 'NON_HANDLED_ERROR'
+    const result = getErrorKind(mockErrorType)
+    expect(result).toEqual(ERROR_KINDS.GENERAL_ERROR)
+  })
+})
