+++ app/src/organisms/ErrorRecoveryFlows/utils/useRouteUpdateActions.ts
@@ -1,88 +1,15 @@
 import * as React from 'react'
-import { useTranslation } from 'react-i18next'
-import head from 'lodash/head'
 import last from 'lodash/last'
+import head from 'lodash/head'
 
-import { RUN_STATUS_AWAITING_RECOVERY } from '@opentrons/api-client'
-import { useCommandQuery } from '@opentrons/react-api-client'
+import { INVALID, RECOVERY_MAP, STEP_ORDER } from '../constants'
 
-import { useNotifyAllCommandsQuery } from '../../resources/runs'
-import { RECOVERY_MAP, ERROR_KINDS, INVALID, STEP_ORDER } from './constants'
-
-import type { RunStatus } from '@opentrons/api-client'
 import type {
-  RouteStep,
   IRecoveryMap,
   RecoveryRoute,
-  ErrorKind,
   RobotMovingRoute,
-  FailedCommand,
-} from './types'
-
-// TODO(jh, 05-09-24): Migrate utils, useRecoveryCommands.ts, and respective tests to a utils dir, and make each util a separate file.
-
-const ALL_COMMANDS_POLL_MS = 5000
-
-// Return the `currentlyRecoveringFrom` command returned by the server, if any.
-// Otherwise, returns null.
-export function useCurrentlyRecoveringFrom(
-  runId: string,
-  runStatus: RunStatus | null
-): FailedCommand | null {
-  // There can only be a currentlyRecoveringFrom command when the run is awaiting-recovery.
-  // In case we're falling back to polling, only enable queries when that is the case.
-  const isRunStatusAwaitingRecovery = runStatus === RUN_STATUS_AWAITING_RECOVERY
-
-  const { data: allCommandsQueryData } = useNotifyAllCommandsQuery(
-    runId,
-    { cursor: null, pageLength: 0 }, // pageLength 0 because we only care about the links.
-    {
-      enabled: isRunStatusAwaitingRecovery,
-      refetchInterval: ALL_COMMANDS_POLL_MS,
-    }
-  )
-  const currentlyRecoveringFromLink =
-    allCommandsQueryData?.links.currentlyRecoveringFrom
-
-  // TODO(mm, 2024-05-21): When the server supports fetching the
-  // currentlyRecoveringFrom command in one step, do that instead of this chained query.
-  const { data: commandQueryData } = useCommandQuery(
-    currentlyRecoveringFromLink?.meta.runId ?? null,
-    currentlyRecoveringFromLink?.meta.commandId ?? null,
-    {
-      enabled:
-        currentlyRecoveringFromLink != null && isRunStatusAwaitingRecovery,
-    }
-  )
-
-  return isRunStatusAwaitingRecovery ? commandQueryData?.data ?? null : null
-}
-
-export function useErrorName(errorKind: ErrorKind): string {
-  const { t } = useTranslation('error_recovery')
-
-  switch (errorKind) {
-    default:
-      return t('general_error')
-  }
-}
-
-// The generalized error message shown to the user in select locations.
-export function useErrorMessage(errorKind: ErrorKind): string {
-  const { t } = useTranslation('error_recovery')
-
-  switch (errorKind) {
-    default:
-      return t('general_error_message')
-  }
-}
-
-export function getErrorKind(errorType?: string): ErrorKind {
-  switch (errorType) {
-    default:
-      return ERROR_KINDS.GENERAL_ERROR
-  }
-}
+  RouteStep,
+} from '../types'
 
 export interface GetRouteUpdateActionsParams {
   hasLaunchedRecovery: boolean
@@ -90,6 +17,7 @@ export interface GetRouteUpdateActionsParams {
   recoveryMap: IRecoveryMap
   setRecoveryMap: (recoveryMap: IRecoveryMap) => void
 }
+
 export interface UseRouteUpdateActionsResult {
   goBackPrevStep: () => Promise<void>
   proceedNextStep: () => Promise<void>
@@ -99,6 +27,7 @@ export interface UseRouteUpdateActionsResult {
     movingRoute?: RobotMovingRoute
   ) => Promise<void>
 }
+
 // Utilities related to routing within the error recovery flows.
 export function useRouteUpdateActions(
   routeUpdateActionsParams: GetRouteUpdateActionsParams
@@ -188,6 +117,7 @@ interface IRecoveryRouteNavigation {
   getNextStep: (step: RouteStep) => RouteStep | typeof INVALID
   getPrevStep: (step: RouteStep) => RouteStep | typeof INVALID
 }
+
 // Returns functions that calculate the next and previous steps of a route given a step.
 export function getRecoveryRouteNavigation(
   route: RecoveryRoute
@@ -223,6 +153,7 @@ interface DetermineRecoveryRoutingParams extends GetRouteUpdateActionsParams {
   updatedStep: string
   currentRoute: RecoveryRoute
 }
+
 // Determine the valid recovery map given the current step.
 // Because RunPausedSplash has access to some ER Wiz routes but is not a part of the ER wizard, the splash
 // is the fallback route as opposed to SelectRecoveryOption (ex, accessed by pressing "go back" enough times).
